Multi-target build: atari + apple2/apple2enh/c64 placeholder ports
- src/main.c is now a thin platform-neutral entry point calling
  platform_run() (declared in src/platform.h, which replaces the unused
  main.h boilerplate)
- the full Atari implementation moved to src/atari/platform.c unchanged
- src/apple2/platform.c and src/c64/platform.c are conio placeholders
  that build and link against their fujinet-lib targets
- Makefile TARGETS = atari apple2 apple2enh c64; all four link
  (fujinet-lib 4.7.4 has release assets for each)
