--- src/main.h
@@ -1,10 +1,0 @@
-#ifndef MAIN_H
-#define MAIN_H
-
-#include <stdint.h>
-#include <stdlib.h>
-
-#include "hello.h"
-#include "util.h"
-
-#endif // MAIN_H
